Replace Linaro builds' description
Linaro builds are not stable for every device and every kernel. It is always better to use the default GCC toolchain to compile builds for a large number of devices. Device maintainers can build Resurrection Remix ROMs using Linaro if they are sure of the stability for their device with Linaro toolchains.